Arreglados numerosos checkstyles
