+++ pom.xml
@@ -18,6 +18,12 @@
             <artifactId>cobertura</artifactId>
             <version>1.9.4.1</version>
         </dependency>
+        <dependency>
+            <groupId>net.java.quickcheck</groupId>
+            <artifactId>quickcheck</artifactId>
+            <version>0.6</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
