+++ src/main/java/servidor/ServRespaldo.java
@@ -1,61 +1,61 @@
+package servidor;
+
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package servidor;
-
 import contenido.Anuncios;
 import contenido.Contenido;
-import utils_exceptions_tokens.InvalidTokenException;
-import servidor_tokens.Token;
+import utils.exceptions.tokens.InvalidTokenException;
+import servidor.tokens.Token;
 import java.util.List;
 
 /**
- * Servidor que implementa un servidor de respaldo para las busquedas
+ * Servidor que implementa un servidor de respaldo para las busquedas.
  *
  * @author Faustino Castro, Victor Blanco y José Miguel del Río
  */
 public class ServRespaldo extends ServLocal {
 
-    private Servidor respaldo;
+  private Servidor respaldo;
 
-    /**
-     * Constructor
-     *
-     * @param nombre
-     */
-    public ServRespaldo(String nombre) {
-        super(nombre);
-    }
+  /**Java Doc.
+   * Constructor.
+   *
+   * @param nombre
+   */
+  public ServRespaldo(String nombre) {
+    super(nombre);
+  }
 
-    /**
-     * Asigna servidor de respaldo
-     *
-     * @param respaldo
-     */
-    public void setServRespaldo(Servidor respaldo) {
-        this.respaldo = respaldo;
-    }
+  /**
+   * Asigna servidor de respaldo.
+   *
+   * @param respaldo
+   */
+  public void setServRespaldo(Servidor respaldo) {
+    this.respaldo = respaldo;
+  }
 
-    /**
-     * Realiza una busqueda entre las canciones almacenadas en el servidor, si
-     * no encuentra ninguna, realizará una búsqueda en el servidor de respaldo.
-     *
-     * @param subcadena
-     * @param token
-     * @return
-     * @throws InvalidTokenException
-     */
-    @Override
-    public List<Contenido> buscar(String subcadena, Token token) throws InvalidTokenException {
-        List<Contenido> lista = super.buscar(subcadena, token);
-        if (lista.isEmpty() || (lista.get(0) instanceof Anuncios && lista.size() == 1)) {
-            respaldo.alta(token);
-            lista = respaldo.buscar(subcadena, token);
-            respaldo.baja(token);
-        }
-        return lista;
+  /**
+   * Realiza una busqueda entre las canciones almacenadas en el servidor, si no
+   * encuentra ninguna, realizará una búsqueda en el servidor de respaldo.
+   *
+   * @param subcadena
+   * @param token
+   * @return
+   * @throws InvalidTokenException
+   */
+  @Override
+  public List<Contenido> buscar(String subcadena, Token token) throws InvalidTokenException {
+    List<Contenido> lista = super.buscar(subcadena, token);
+    if (lista.isEmpty() || (lista.get(0) instanceof Anuncios && lista.size() == 1)) {
+      respaldo.alta(token);
+      lista = respaldo.buscar(subcadena, token);
+      respaldo.baja(token);
     }
+    return lista;
+  }
 
 }
