+++ src/test/java/servidor/test/ServRespaldoTest.java
@@ -1,10 +1,10 @@
+package servidor.test;
+
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package servidor_test;
-
 import servidor.ServRespaldo;
 import org.junit.After;
 import org.junit.AfterClass;
@@ -14,63 +14,63 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 
 /**
- *
+ *Java Doc.
  * @author Faustino Castro, Victor Blanco y José Miguel del Río
  */
 public class ServRespaldoTest {
 
-    /**
-     *
-     */
-    public ServRespaldoTest() {
-    }
+  /**
+   *Java Doc.
+   */
+  public ServRespaldoTest() {
+  }
 
-    /**
-     *
-     */
-    @BeforeClass
-    public static void setUpClass() {
-    }
+  /**
+   *Java Doc.
+   */
+  @BeforeClass
+  public static void setUpClass() {
+  }
 
-    /**
-     *
-     */
-    @AfterClass
-    public static void tearDownClass() {
-    }
+  /**
+   *Java Doc.
+   */
+  @AfterClass
+  public static void tearDownClass() {
+  }
 
-    /**
-     *
-     */
-    @Before
-    public void setUp() {
-    }
+  /**
+   *Java Doc.
+   */
+  @Before
+  public void setUp() {
+  }
 
-    /**
-     *
-     */
-    @After
-    public void tearDown() {
-    }
+  /**
+   *Java Doc.
+   */
+  @After
+  public void tearDown() {
+  }
 
-    /**
-     * Test of setServRespaldo method, of class ServRespaldo. crear un servidor
-     * de respaldo y obtener su nombre
-     */
-    @Test
-    public void testSetServRespaldo() {
-        ServRespaldo servidor = new ServRespaldo("Torre");
-        servidor.setServRespaldo(servidor);
+  /**
+   * Test of setServRespaldo method, of class ServRespaldo. crear un servidor de.
+   * respaldo y obtener su nombre
+   */
+  @Test
+  public void testSetServRespaldo() {
+    ServRespaldo servidor = new ServRespaldo("Torre");
+    servidor.setServRespaldo(servidor);
 
-        String nombre = servidor.obtenerNombre();
+    String nombre = servidor.obtenerNombre();
 
-        assertEquals(nombre, "Torre");
+    assertEquals(nombre, "Torre");
 
-    }
+  }
 
-    /**
-     * Test of buscar method, of class ServRespaldo.
-     */
+  /**
+   * Test of buscar method, of class ServRespaldo.
+   */
 //    @Test
 //    public void testBuscar() throws Exception {
 //        ServRespaldo servidor = new ServRespaldo("Torre");
